Fixed for new Rev firmware(1.8.2)
Also had some problems where we weren't initializing the variables each time we were at a target (to use for the next target rotation and drive).

Per Mr. Harsch's suggestion we are rotating a full 360 deg when we can't find a target.

Currently during this 360 deg rotation we aren't actually doing a full rotation so we are ending up facing the wrong direction...  Next version will go back to one rotation